+++ src/xgboost_distribution/model.py
@@ -137,6 +137,12 @@ class XGBDistribution(XGBModel, RegressorMixin):
         else:
             base_margin_eval_set = None
 
+        for param in [sample_weight, sample_weight_eval_set]:
+            if param is not None:
+                raise NotImplementedError(
+                    "Sample weights are currently not supported by XGBDistribution!"
+                )
+
         train_dmatrix, evals = _wrap_evaluation_matrices(
             missing=self.missing,
             X=X,